requirements.txt
scipy
numpy
opencv-python
flask
Pillow
keras
tensorflow
scikit-learn
